+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Support for a custom "range" option to build a BVH from a sub range of the geometry.
+
 ## [0.7.6] - 2024-07-02
 ### Fixed
 - Regression when intersecting InstancedMesh instances.